requirements.txt
gradio==4.44.0
numpy==1.24.3
matplotlib==3.7.2
pillow==10.0.0
